# recommendation/src/main/java/com/example/recommendation/dtos/AssessmentTriggerDTO.java
package com.example.recommendation.dtos;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Data;

@Data
public class AssessmentTriggerDTO {
    @Schema(description = "ID do usuário")
    private Long userId;
    
    @Schema(description = "ID da competência")
    private Long competencyId;
}